requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
numpy==2.3.5
packaging==25.0
pandas==2.2.3
plotly==5.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
six==1.17.0
tenacity==9.1.2
tzdata==2025.2
urllib3==2.6.1
yara-python==4.5.4
